dom/workers/test/serviceworkers/test_close.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1131353 - test WorkerGlobalScope.close() on service workers</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
<iframe></iframe>
</div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  var iframe;
  function runTest() {
    navigator.serviceWorker.ready.then(setupSW);
    navigator.serviceWorker.register("close_test.js", {scope: "."});

    function setupSW(registration) {
      var worker = registration.waiting ||
                   registration.active;
      var iframe = document.createElement("iframe");
      iframe.src = "message_receiver.html";
      iframe.onload = function() {
        worker.postMessage({ message: "start" });
      };
      document.body.appendChild(iframe);
    }

    window.onmessage = function(e) {
      if (e.data.status == "ok") {
        ok(e.data.result, e.data.message);
      } else if (e.data.status == "done") {
        navigator.serviceWorker.getRegistration().then(function(registration) {
          registration.unregister().then(function(result) {
            ok(result, "Unregistering the service worker should succeed");
            SimpleTest.finish();
          }, function(e) {
            dump("Unregistering the SW failed with " + e + "\n");
            SimpleTest.finish();
          });
        });
      }
    };
  }

  SimpleTest.waitForExplicitFinish();
  onload = function() {
    SpecialPowers.pushPrefEnv({"set": [
      ["dom.serviceWorkers.exemptFromPerDomainMax", true],
      ["dom.serviceWorkers.enabled", true],
      ["dom.serviceWorkers.testing.enabled", true],
    ]}, runTest);
  };
</script>
</pre>
</body>
</html>
